+++ CHANGES.rst
@@ -12,6 +12,10 @@ Version 0.18
 Unreleased
 
 * Support Python 3.8.
+* Increase the :meth:`~Reader.get_entries` recent threshold from 3 to 7 days.
+  (`#141`_)
+
+.. _#141: https://github.com/lemon24/reader/issues/141
 
 
 Version 0.17